Rename and update shift to helpable
